Rework test_skip_input_fifo() and test_seek_output_fifo() to not use the system provided dd utility
On BSD and macOS, the system provided dd utility opens up the output
file for both read and write. This means that the write to the fifo does
not block, and instantly completes. This leaves nothing left to read by
the time the coreutils-rs dd tries to read the fifo.